decompose logic from getUsersForStatistic to helper functions
