+++ modules/user/user.service.js
@@ -18,6 +18,11 @@ const {
   adminConfirmationMessage,
 } = require('../../utils/localization');
 const { uploadFiles, deleteFiles } = require('../upload/upload.service');
+const {
+  removeDaysFromData,
+  countItemsCreatedAtUniqueDate,
+  changeItemToUniqueDate,
+} = require('../helper-functions');
 
 const {
   USER_ALREADY_EXIST,
@@ -55,15 +60,6 @@ class UserService {
         $gte: removeDaysFromData(days, Date.now()),
         $lte: removeDaysFromData(0, Date.now()),
       };
-    }
-
-    function removeDaysFromData(days, currentDate) {
-      const currentDay = new Date(Date.now()).getDate();
-      const timeFewDaysAgo = new Date(currentDate).setDate(currentDay - days);
-      return `
-        ${new Date(timeFewDaysAgo).getFullYear()}-${new Date(
-        timeFewDaysAgo
-      ).getMonth() + 1}-${new Date(timeFewDaysAgo).getDate()}`;
     }
 
     return filter;
@@ -104,68 +100,9 @@ class UserService {
   async getUsersForStatistic({ filter }) {
     const filters = this.filterItems(filter);
     const items = await User.find(filters);
-    const labels = Array.from(
-      new Set(
-        items
-          .map(
-            el =>
-              `${new Date(el.registrationDate).getDate()} ${monthNumToStr(
-                new Date(el.registrationDate).getMonth()
-              )}`
-          )
-          .sort((a, b) => {
-            a.split('.')[0] - b;
-          })
-      )
-    );
+    const labels = changeItemToUniqueDate(items);
 
-    const data = Object.values(
-      items
-        .map(
-          el =>
-            `${new Date(el.registrationDate).getDate()}.${new Date(
-              el.registrationDate.getMonth()
-            )}`
-        )
-        .sort((a, b) => {
-          a.split('.')[0] - b;
-        })
-        .reduce((acc, el) => {
-          acc[el] = (acc[el] || 0) + 1;
-          return acc;
-        }, {})
-    );
-
-    function monthNumToStr(num) {
-      switch (+num) {
-        case 0:
-          return 'Jan';
-        case 1:
-          return 'Feb';
-        case 2:
-          return 'Mar';
-        case 3:
-          return 'Apr';
-        case 4:
-          return 'May';
-        case 5:
-          return 'Jun';
-        case 6:
-          return 'Jul';
-        case 7:
-          return 'Aug';
-        case 8:
-          return 'Sep';
-        case 9:
-          return 'Oct';
-        case 10:
-          return 'Nov';
-        case 11:
-          return 'Dec';
-        default:
-          return '';
-      }
-    }
+    const data = countItemsCreatedAtUniqueDate(items);
 
     console.log(labels, data);
     return { labels, data };
